FRI-14 Moved security config to DispatcherServletContext.xml to get spring security working, and update logic for create/update product

+++ api/src/main/java/org/ihtsdo/buildcloud/controller/BuildController.java
@@ -79,6 +79,10 @@ public class BuildController {
 
 		ifBuildIsNullThrow(productKey, buildId, build);
 
+		if (build.getTags().contains(Build.Tag.PUBLISHED)) {
+			throw new BusinessServiceException("You can not delete the PUBLISHED build");
+		}
+
 		buildService.delete(releaseCenterKey, productKey, buildId);
 
 		return new ResponseEntity<>(HttpStatus.OK);
